+++ dropwizard-guicey/src/main/java/ru/vyarus/dropwizard/guice/test/builder/TestSupportBuilder.java
@@ -1,29 +1,19 @@
-package ru.vyarus.dropwizard.guice.test.util.support;
+package ru.vyarus.dropwizard.guice.test.builder;
 
 import com.google.common.base.MoreObjects;
-import com.google.common.base.Preconditions;
 import com.google.inject.Injector;
-import io.dropwizard.configuration.ConfigurationSourceProvider;
 import io.dropwizard.Application;
 import io.dropwizard.Configuration;
-import io.dropwizard.testing.ConfigOverride;
 import io.dropwizard.testing.DropwizardTestSupport;
 import javax.annotation.Nullable;
-import ru.vyarus.dropwizard.guice.hook.GuiceyConfigurationHook;
 import ru.vyarus.dropwizard.guice.test.GuiceyTestSupport;
 import ru.vyarus.dropwizard.guice.test.TestSupport;
-import ru.vyarus.dropwizard.guice.test.util.ConfigOverrideUtils;
-import ru.vyarus.dropwizard.guice.test.util.HooksUtil;
 import ru.vyarus.dropwizard.guice.test.util.RandomPortsListener;
-import ru.vyarus.dropwizard.guice.test.util.client.DefaultTestClientFactory;
-import ru.vyarus.dropwizard.guice.test.util.client.TestClientFactory;
+import ru.vyarus.dropwizard.guice.test.client.DefaultTestClientFactory;
+import ru.vyarus.dropwizard.guice.test.client.TestClientFactory;
 
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.function.Supplier;
 
 /**
  * Builder and runner for {@link io.dropwizard.testing.DropwizardTestSupport} and
@@ -31,127 +21,24 @@ import java.util.function.Supplier;
  * should be suitable for cases when junit 5 extensions could not be used.
  * <p>
  * Use {@link ru.vyarus.dropwizard.guice.test.TestSupport#build(Class)} to build instance.
+ * <p>
+ * Builder is not supposed to be used for multiple runs: registered hooks will be applied only once. This limitation
+ * is not possible to avoid because builder could be used for support objects creation, which are not aware of
+ * hooks. So hooks could be registered globally only in time of addition to the builder.
  *
  * @param <C> configuration type
  * @author Vyacheslav Rusakov
  * @since 14.11.2023
  */
-@SuppressWarnings({"PMD.TooManyMethods", "PMD.GodClass", "PMD.AvoidFieldNameMatchingMethodName"})
-public class TestSupportBuilder<C extends Configuration> {
+@SuppressWarnings("PMD.AvoidFieldNameMatchingMethodName")
+public class TestSupportBuilder<C extends Configuration> extends BaseBuilder<C, TestSupportBuilder<C>> {
 
-    private final Class<? extends Application<C>> app;
-    private String configPath;
-    private ConfigurationSourceProvider configSourceProvider;
-    private final Map<String, Supplier<String>> configOverrides = new HashMap<>();
-    private C configObject;
-    private String propertyPrefix;
     private boolean randomPorts;
-    private String restMapping;
     private final List<TestListener<C>> listeners = new ArrayList<>();
     private TestClientFactory factory = new DefaultTestClientFactory();
 
     public TestSupportBuilder(final Class<? extends Application<C>> app) {
-        this.app = app;
-    }
-
-    /**
-     * Must not be used if {@link #config(io.dropwizard.core.Configuration)} used.
-     *
-     * @param path configuration file path
-     * @return builder instance for chained calls
-     */
-    public TestSupportBuilder<C> config(final @Nullable String path) {
-        this.configPath = path;
-        return this;
-    }
-
-    /**
-     * Use configuration instance instead of configuration parsing from yaml file. When this is used, other
-     * configuration options must not be used (they can't be used, and an error would be thrown indicating incorrect
-     * usage).
-     *
-     * @param config pre-initialized configuration object
-     * @return builder instance for chained calls
-     */
-    public TestSupportBuilder<C> config(final @Nullable C config) {
-        this.configObject = config;
-        return this;
-    }
-
-    /**
-     * Must not be used if {@link #config(io.dropwizard.core.Configuration)} used.
-     *
-     * @param provider configuration source provider
-     * @return builder instance for chained calls
-     */
-    public TestSupportBuilder<C> configSourceProvider(final @Nullable ConfigurationSourceProvider provider) {
-        this.configSourceProvider = provider;
-        return this;
-    }
-
-    /**
-     * Must not be used if {@link #config(io.dropwizard.core.Configuration)} used.
-     *
-     * @param overrides config override values (in format "path: value")
-     * @return builder instance for chained calls
-     */
-    public TestSupportBuilder<C> configOverrides(final String... overrides) {
-        for (String over : overrides) {
-            configOverride(over);
-        }
-        return this;
-    }
-
-    /**
-     * Must not be used if {@link #config(io.dropwizard.core.Configuration)} used.
-     *
-     * @param override config override value (in format "path: value")
-     * @return builder instance for chained calls
-     */
-    public TestSupportBuilder<C> configOverride(final @Nullable String override) {
-        if (override != null) {
-            final int idx = override.indexOf(':');
-            Preconditions.checkState(idx > 0,
-                    "Incorrect configuration override declaration: must be 'key: value', but found '%s'", override);
-            configOverride(override.substring(0, idx).trim(), override.substring(idx + 1).trim());
-        }
-        return this;
-    }
-
-    /**
-     * Must not be used if {@link #config(io.dropwizard.core.Configuration)} used.
-     *
-     * @param key   configuration path
-     * @param value overriding value
-     * @return builder instance for chained calls
-     */
-    public TestSupportBuilder<C> configOverride(final String key, final String value) {
-        return configOverride(key, () -> value);
-    }
-
-    /**
-     * Must not be used if {@link #config(io.dropwizard.core.Configuration)} used.
-     *
-     * @param key   configuration path
-     * @param value overriding value provider
-     * @return builder instance for chained calls
-     */
-    public TestSupportBuilder<C> configOverride(final String key, final Supplier<String> value) {
-        this.configOverrides.put(key, value);
-        return this;
-    }
-
-    /**
-     * Dropwizard stored all provided configuration overriding values as system properties with provided prefix
-     * (or "dw." by default). If multiple tests run concurrently, they would collide on using the same system
-     * properties. It is preferred to specify test-unique prefix.
-     *
-     * @param prefix configuration override properties prefix
-     * @return builder instance for chained calls
-     */
-    public TestSupportBuilder<C> propertyPrefix(final @Nullable String prefix) {
-        this.propertyPrefix = prefix;
-        return this;
+        super(app);
     }
 
     /**
@@ -170,23 +57,6 @@ public class TestSupportBuilder<C extends Configuration> {
      */
     public TestSupportBuilder<C> randomPorts(final boolean randomPorts) {
         this.randomPorts = randomPorts;
-        return this;
-    }
-
-    /**
-     * Specifies rest mapping path. This is the same as specifying direct config override
-     * {@code "server.rootMapping: /something/*"}. Specified value would be prefixed with "/" and, if required
-     * "/*" applied at the end. So it would be correct to specify {@code restMapping = "api"} (actually set value
-     * would be "/api/*").
-     * <p>
-     * This option is only intended to simplify cases when custom configuration file is not yet used in tests
-     * (usually early PoC phase). It allows you to map servlet into application root in test (because rest is no
-     * more resided in root). When used with existing configuration file, this parameter will override file definition.
-     *
-     * @return builder instance for chained calls
-     */
-    public TestSupportBuilder<C> restMapping(final String restMapping) {
-        this.restMapping = restMapping;
         return this;
     }
 
@@ -214,29 +84,6 @@ public class TestSupportBuilder<C extends Configuration> {
      */
     public TestSupportBuilder<C> listen(final TestListener<C> listener) {
         this.listeners.add(listener);
-        return this;
-    }
-
-    /**
-     * Shortcut for hooks registration (method simply immediately registers provided hooks).
-     *
-     * @param hooks hook classes to install (nulls not allowed)
-     * @return builder instance for chained calls
-     */
-    @SafeVarargs
-    public final TestSupportBuilder<C> hooks(final Class<? extends GuiceyConfigurationHook>... hooks) {
-        HooksUtil.register(HooksUtil.create(hooks));
-        return this;
-    }
-
-    /**
-     * Shortcut for hooks registration (method simply immediately registers provided hooks).
-     *
-     * @param hooks hooks to install (nulls allowed)
-     * @return builder instance for chained calls
-     */
-    public TestSupportBuilder<C> hooks(final GuiceyConfigurationHook... hooks) {
-        HooksUtil.register(Arrays.asList(hooks));
         return this;
     }
 
@@ -277,6 +124,9 @@ public class TestSupportBuilder<C extends Configuration> {
     /**
      * Start and stop application without web services. Mostly useful to test application startup errors
      * (with proper application shutdown).
+     * <p>
+     * NOTE: method not supposed to be used for multiple calls. For example, registered hooks would only work
+     * on first execution.
      *
      * @return test support object used for execution (for example, to examine configuration)
      * @throws Exception any appeared exception (throws may easily be added directly to test method and, without
@@ -288,6 +138,9 @@ public class TestSupportBuilder<C extends Configuration> {
 
     /**
      * Start and stop application without web services. Provided action would be executed in time of application life.
+     * <p>
+     * NOTE: method not supposed to be used for multiple calls. For example, registered hooks would only work
+     * on first execution.
      *
      * @param action action to execute while the application is running
      * @param <T>    result type
@@ -302,6 +155,9 @@ public class TestSupportBuilder<C extends Configuration> {
     /**
      * Start and stop application with web services. Mostly useful to test application startup errors
      * (with proper application shutdown).
+     * <p>
+     * NOTE: method not supposed to be used for multiple calls. For example, registered hooks would only work
+     * on first execution.
      *
      * @return test support object used for execution (for example, to examine configuration)
      * @throws Exception any appeared exception (throws may easily be added directly to test method and, without
@@ -313,6 +169,9 @@ public class TestSupportBuilder<C extends Configuration> {
 
     /**
      * Start and stop application with web services. Provided action would be executed in time of application life.
+     * <p>
+     * NOTE: method not supposed to be used for multiple calls. For example, registered hooks would only work
+     * on first execution.
      *
      * @param action action to execute while the application is running
      * @param <T>    result type
@@ -361,19 +220,6 @@ public class TestSupportBuilder<C extends Configuration> {
         }
 
         return support;
-    }
-
-    private ConfigOverride[] prepareOverrides(final String prefix) {
-        final ConfigOverride[] override = new ConfigOverride[configOverrides.size() + (restMapping == null ? 0 : 1)];
-        int i = 0;
-        for (Map.Entry<String, Supplier<String>> entry : configOverrides.entrySet()) {
-            override[i++] = ConfigOverride.config(prefix,
-                    entry.getKey(), entry.getValue());
-        }
-        if (restMapping != null) {
-            override[i] = ConfigOverrideUtils.overrideRestMapping(prefix, restMapping);
-        }
-        return override;
     }
 
     @SuppressWarnings("PMD.AvoidRethrowingException")