+++ dropwizard-guicey/src/main/java/ru/vyarus/dropwizard/guice/test/jupiter/TestGuiceyApp.java
@@ -6,8 +6,8 @@ import ru.vyarus.dropwizard.guice.hook.GuiceyConfigurationHook;
 import ru.vyarus.dropwizard.guice.test.ClientSupport;
 import ru.vyarus.dropwizard.guice.test.jupiter.env.TestEnvironmentSetup;
 import ru.vyarus.dropwizard.guice.test.jupiter.ext.TestGuiceyAppExtension;
-import ru.vyarus.dropwizard.guice.test.util.client.DefaultTestClientFactory;
-import ru.vyarus.dropwizard.guice.test.util.client.TestClientFactory;
+import ru.vyarus.dropwizard.guice.test.client.DefaultTestClientFactory;
+import ru.vyarus.dropwizard.guice.test.client.TestClientFactory;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Inherited;